refactor(annotations): move the rowid resolvers into the annotations helper
Fold entity_rowids into helpers/annotations alongside the other note
helpers, and take the expected node type names from the GraphQL classes
instead of string literals.
